# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proector API",
        "description": "Backend API documentation for Proector mobile and admin platforms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        },
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/security/orders": {
            "get": {
                "tags": [
                    "Security Personnel Orders"
                ],
                "summary": "Get active orders for security personnel",
                "description": "Returns active bookings assigned to authenticated security personnel team.",
                "operationId": "4facf5590084703e1fd43d53a5551056",
                "responses": {
                    "200": {
                        "description": "Orders list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/security/orders/history": {
            "get": {
                "tags": [
                    "Security Personnel Orders"
                ],
                "summary": "Get completed order history",
                "description": "Returns completed and cancelled orders assigned to security team.",
                "operationId": "82e7fe0e403d3a97c1a2dbee816ebdff",
                "responses": {
                    "200": {
                        "description": "Order history list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/security/orders/{id}": {
            "get": {
                "tags": [
                    "Security Personnel Orders"
                ],
                "summary": "Get order details",
                "description": "Returns details of a specific assigned order including chat and latest location.",
                "operationId": "c77612c4aa6dee92434bf7e018096aa1",
                "responses": {
                    "200": {
                        "description": "Order details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/security/orders/{id}/en-route": {
            "post": {
                "tags": [
                    "Security Personnel Status"
                ],
                "summary": "Mark order as en route",
                "description": "Transitions booking to ongoing state and starts live location tracking.",
                "operationId": "6f5739f6d3fba488b7c091d3524fe747",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/security/orders/{id}/arrived": {
            "post": {
                "tags": [
                    "Security Personnel Status"
                ],
                "summary": "Mark order as arrived",
                "description": "Transitions booking to arrived state and records latest team location.",
                "operationId": "c9502b11546766ac4c0c9b86d5614c1b",
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/security/orders/{id}/complete": {
            "post": {
                "tags": [
                    "Security Personnel Status"
                ],
                "summary": "Mark order as completed",
                "description": "Transitions arrived booking to completed and frees team/personnel availability.",
                "operationId": "81b5693565ed1cba99b038ab8219e712",
                "responses": {
                    "200": {
                        "description": "Booking completed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/security/location/update": {
            "post": {
                "tags": [
                    "Security Personnel Status"
                ],
                "summary": "Update location tracking",
                "description": "Pushes periodic location updates for active booking tracking.",
                "operationId": "f8d315478f4562e82eeda9a2f3d5374f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_id",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "booking_id": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    },
                                    "accuracy": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "speed": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "heading": {
                                        "type": "number",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Bearer {token} from login or OTP verification endpoints.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}